sales transaction shared component and to be opened from sales order
